requirements.txt
torch==2.2.2
torchaudio
torchvision
transformers==4.41.0
datasets
bitsandbytes
python-dotenv
discord.py
peft
trl
setuptools